Move conditional order notifications to common package
