+++ src/main/java/uk/gov/hmcts/divorce/common/notification/AwaitingConditionalOrderReminderNotification.java
@@ -1,22 +1,26 @@
-package uk.gov.hmcts.divorce.notification;
+package uk.gov.hmcts.divorce.common.notification;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.notification.ApplicantNotification;
+import uk.gov.hmcts.divorce.notification.CommonContent;
+import uk.gov.hmcts.divorce.notification.NotificationService;
 
 import java.util.Map;
 
 import static java.util.Objects.nonNull;
 import static uk.gov.hmcts.divorce.notification.CommonContent.IS_REMINDER;
 import static uk.gov.hmcts.divorce.notification.CommonContent.SOLICITOR_NAME;
+import static uk.gov.hmcts.divorce.notification.CommonContent.YES;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.CITIZEN_APPLY_FOR_CONDITIONAL_ORDER;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLICITOR_AWAITING_CONDITIONAL_ORDER;
 
 @Component
 @Slf4j
-public class AwaitingConditionalOrderNotification implements ApplicantNotification {
+public class AwaitingConditionalOrderReminderNotification implements ApplicantNotification {
 
     @Autowired
     private CommonContent commonContent;
@@ -26,11 +30,13 @@ public class AwaitingConditionalOrderNotification implements ApplicantNotificati
 
     @Override
     public void sendToApplicant1(final CaseData caseData, final Long id) {
-        log.info("Notifying applicant 1 that they can apply for a conditional order: {}", id);
+        log.info("Sending reminder to applicant 1 that they can apply for a conditional order: {}", id);
+
         final Applicant applicant1 = caseData.getApplicant1();
+
         final Map<String, String> templateVars = commonContent
             .conditionalOrderTemplateVars(caseData, id, applicant1, caseData.getApplicant2());
-        templateVars.put(IS_REMINDER, CommonContent.NO);
+        templateVars.put(IS_REMINDER, YES);
 
         notificationService.sendEmail(
             applicant1.getEmail(),
@@ -42,9 +48,9 @@ public class AwaitingConditionalOrderNotification implements ApplicantNotificati
 
     @Override
     public void sendToApplicant1Solicitor(final CaseData caseData, final Long id) {
-        log.info("Notifying applicant 1 solicitor that they can apply for a conditional order: {}", id);
-
+        log.info("Sending reminder to applicant 1 solicitor that they can apply for a conditional order: {}", id);
         final Applicant applicant1 = caseData.getApplicant1();
+
         final Map<String, String> templateVars = commonContent.basicTemplateVars(caseData, id);
         templateVars.put(SOLICITOR_NAME, applicant1.getSolicitor().getName());
 
@@ -54,17 +60,18 @@ public class AwaitingConditionalOrderNotification implements ApplicantNotificati
             templateVars,
             applicant1.getLanguagePreference()
         );
-        log.info("Successfully sent awaiting conditional order notification for case : {}", id);
     }
 
     @Override
     public void sendToApplicant2(final CaseData caseData, final Long id) {
         if (!caseData.getApplicationType().isSole() && nonNull(caseData.getApplicant2().getEmail())) {
-            log.info("Notifying applicant 2 that they can apply for a conditional order: {}", id);
+            log.info("Sending reminder applicant 2 that they can apply for a conditional order: {}", id);
+
             final Applicant applicant2 = caseData.getApplicant2();
+
             final Map<String, String> templateVars = commonContent
                 .conditionalOrderTemplateVars(caseData, id, applicant2, caseData.getApplicant1());
-            templateVars.put(IS_REMINDER, CommonContent.NO);
+            templateVars.put(IS_REMINDER, YES);
 
             notificationService.sendEmail(
                 applicant2.getEmail(),
@@ -78,9 +85,9 @@ public class AwaitingConditionalOrderNotification implements ApplicantNotificati
     @Override
     public void sendToApplicant2Solicitor(final CaseData caseData, final Long id) {
         if (!caseData.getApplicationType().isSole()) {
-            log.info("Notifying applicant 2 solicitor that they can apply for a conditional order: {}", id);
-
+            log.info("Sending reminder applicant 2 solicitor that they can apply for a conditional order: {}", id);
             final Applicant applicant2 = caseData.getApplicant2();
+
             final Map<String, String> templateVars = commonContent.basicTemplateVars(caseData, id);
             templateVars.put(SOLICITOR_NAME, applicant2.getSolicitor().getName());
