+++ src/main/java/uk/gov/hmcts/divorce/common/notification/ConditionalOrderPendingReminderNotification.java
@@ -1,10 +1,13 @@
-package uk.gov.hmcts.divorce.notification;
+package uk.gov.hmcts.divorce.common.notification;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.notification.ApplicantNotification;
+import uk.gov.hmcts.divorce.notification.CommonContent;
+import uk.gov.hmcts.divorce.notification.NotificationService;
 
 import java.util.Map;
