+++ src/test/java/uk/gov/hmcts/divorce/common/notification/AwaitingConditionalOrderReminderNotificationTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.divorce.notification;
+package uk.gov.hmcts.divorce.common.notification;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -6,10 +6,13 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
+import uk.gov.hmcts.divorce.common.notification.AwaitingConditionalOrderReminderNotification;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder;
 import uk.gov.hmcts.divorce.divorcecase.model.Solicitor;
+import uk.gov.hmcts.divorce.notification.CommonContent;
+import uk.gov.hmcts.divorce.notification.NotificationService;
 
 import java.util.HashMap;
 
@@ -20,6 +23,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static org.mockito.hamcrest.MockitoHamcrest.argThat;
+import static uk.gov.hmcts.divorce.divorcecase.model.ApplicationType.JOINT_APPLICATION;
+import static uk.gov.hmcts.divorce.divorcecase.model.ApplicationType.SOLE_APPLICATION;
 import static uk.gov.hmcts.divorce.divorcecase.model.LanguagePreference.ENGLISH;
 import static uk.gov.hmcts.divorce.notification.CommonContent.APPLICANT_NAME;
 import static uk.gov.hmcts.divorce.notification.CommonContent.APPLICATION_REFERENCE;
@@ -42,7 +47,7 @@ import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getBasicTemplateVars;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getConditionalOrderQuestions;
 
 @ExtendWith(MockitoExtension.class)
-class ConditionalOrderPendingReminderNotificationTest {
+class AwaitingConditionalOrderReminderNotificationTest {
 
     @Mock
     private CommonContent commonContent;
@@ -51,15 +56,11 @@ class ConditionalOrderPendingReminderNotificationTest {
     private NotificationService notificationService;
 
     @InjectMocks
-    private ConditionalOrderPendingReminderNotification conditionalOrderPendingReminderNotification;
+    private AwaitingConditionalOrderReminderNotification awaitingConditionalOrderReminderNotification;
 
     @Test
-    void shouldSendNotificationToApplicant1WhenSubmittedDateIsNotSet() {
+    void shouldSendNotificationToApplicant1() {
         final CaseData caseData = caseData();
-        caseData.setConditionalOrder(ConditionalOrder.builder()
-            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
-            .build());
-        caseData.getConditionalOrder().getConditionalOrderApplicant1Questions().setSubmittedDate(null);
 
         when(commonContent
             .conditionalOrderTemplateVars(
@@ -69,7 +70,7 @@ class ConditionalOrderPendingReminderNotificationTest {
                 caseData.getApplicant2()))
             .thenReturn(new HashMap<>());
 
-        conditionalOrderPendingReminderNotification.sendToApplicant1(caseData, 1234567890123456L);
+        awaitingConditionalOrderReminderNotification.sendToApplicant1(caseData, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
@@ -82,19 +83,7 @@ class ConditionalOrderPendingReminderNotificationTest {
     }
 
     @Test
-    void shouldNotSendNotificationToApplicant1WhenSubmittedDateIsSet() {
-        final CaseData caseData = caseData();
-        caseData.setConditionalOrder(ConditionalOrder.builder()
-            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
-            .build());
-
-        conditionalOrderPendingReminderNotification.sendToApplicant1(caseData, 1234567890123456L);
-
-        verifyNoInteractions(notificationService);
-    }
-
-    @Test
-    void shouldSendNotificationToApplicant1SolicitorWhenSubmittedDateIsNotSet() {
+    void shouldSendNotificationToApplicant1Solicitor() {
         final Applicant applicant = getApplicant();
         applicant.setSolicitor(Solicitor.builder().email(TEST_SOLICITOR_EMAIL).name(TEST_SOLICITOR_NAME).build());
         applicant.setSolicitorRepresented(YesOrNo.YES);
@@ -102,11 +91,10 @@ class ConditionalOrderPendingReminderNotificationTest {
         caseData.setConditionalOrder(ConditionalOrder.builder()
             .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
             .build());
-        caseData.getConditionalOrder().getConditionalOrderApplicant1Questions().setSubmittedDate(null);
 
         when(commonContent.basicTemplateVars(caseData, 1234567890123456L)).thenReturn(getBasicTemplateVars());
 
-        conditionalOrderPendingReminderNotification.sendToApplicant1Solicitor(caseData, 1234567890123456L);
+        awaitingConditionalOrderReminderNotification.sendToApplicant1Solicitor(caseData, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_SOLICITOR_EMAIL),
@@ -122,27 +110,10 @@ class ConditionalOrderPendingReminderNotificationTest {
     }
 
     @Test
-    void shouldNotSendNotificationToApplicant1SolicitorWhenSubmittedDateIsSet() {
-        final Applicant applicant = getApplicant();
-        applicant.setSolicitor(Solicitor.builder().email(TEST_SOLICITOR_EMAIL).name(TEST_SOLICITOR_NAME).build());
-        applicant.setSolicitorRepresented(YesOrNo.YES);
-        final var caseData = CaseData.builder().applicant1(applicant).build();
-        caseData.setConditionalOrder(ConditionalOrder.builder()
-            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
-            .build());
-
-        conditionalOrderPendingReminderNotification.sendToApplicant1Solicitor(caseData, 1234567890123456L);
-
-        verifyNoInteractions(notificationService);
-    }
-
-    @Test
-    void shouldSendNotificationToApplicant2WhenSubmittedDateIsSet() {
+    void shouldSendNotificationToApplicant2IfJointApplication() {
         final CaseData caseData = caseData();
+        caseData.setApplicationType(JOINT_APPLICATION);
         caseData.getApplicant2().setEmail(TEST_APPLICANT_2_USER_EMAIL);
-        caseData.setConditionalOrder(ConditionalOrder.builder()
-            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
-            .build());
 
         when(commonContent
             .conditionalOrderTemplateVars(
@@ -152,7 +123,7 @@ class ConditionalOrderPendingReminderNotificationTest {
                 caseData.getApplicant1()))
             .thenReturn(new HashMap<>());
 
-        conditionalOrderPendingReminderNotification.sendToApplicant2(caseData, 1234567890123456L);
+        awaitingConditionalOrderReminderNotification.sendToApplicant2(caseData, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_APPLICANT_2_USER_EMAIL),
@@ -165,61 +136,30 @@ class ConditionalOrderPendingReminderNotificationTest {
     }
 
     @Test
-    void shouldNotSendNotificationToApplicant2WhenSubmittedDateIsNotSet() {
+    void shouldNotSendNotificationToApplicant2IfSoleApplication() {
         final CaseData caseData = caseData();
+        caseData.setApplicationType(SOLE_APPLICATION);
         caseData.getApplicant2().setEmail(TEST_APPLICANT_2_USER_EMAIL);
-        caseData.setConditionalOrder(ConditionalOrder.builder()
-            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
-            .build());
-        caseData.getConditionalOrder().getConditionalOrderApplicant1Questions().setSubmittedDate(null);
 
-        conditionalOrderPendingReminderNotification.sendToApplicant2(caseData, 1234567890123456L);
+        awaitingConditionalOrderReminderNotification.sendToApplicant2(caseData, 1234567890123456L);
 
         verifyNoInteractions(notificationService);
     }
 
     @Test
-    void shouldNotSendNotificationToApplicant2WhenApplicant2EmailNotSet() {
-        final CaseData caseData = caseData();
-        caseData.getApplicant2().setEmail(null);
-        caseData.setConditionalOrder(ConditionalOrder.builder()
-            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
-            .build());
-
-        conditionalOrderPendingReminderNotification.sendToApplicant2(caseData, 1234567890123456L);
-
-        verifyNoInteractions(notificationService);
-    }
-
-    @Test
-    void shouldNotSendNotificationToApplicant2SolicitorWhenSubmittedDateIsNotSet() {
+    void shouldSendNotificationToApplicant2SolicitorIfJointApplication() {
         final Applicant applicant = getApplicant();
         applicant.setSolicitor(Solicitor.builder().email(TEST_SOLICITOR_EMAIL).name(TEST_SOLICITOR_NAME).build());
         applicant.setSolicitorRepresented(YesOrNo.YES);
         final var caseData = CaseData.builder().applicant2(applicant).build();
-        caseData.setConditionalOrder(ConditionalOrder.builder()
-            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
-            .build());
-        caseData.getConditionalOrder().getConditionalOrderApplicant1Questions().setSubmittedDate(null);
-
-        conditionalOrderPendingReminderNotification.sendToApplicant2Solicitor(caseData, 1234567890123456L);
-
-        verifyNoInteractions(notificationService);
-    }
-
-    @Test
-    void shouldSendNotificationToApplicant2SolicitorWhenSubmittedDateIsSet() {
-        final Applicant applicant = getApplicant();
-        applicant.setSolicitor(Solicitor.builder().email(TEST_SOLICITOR_EMAIL).name(TEST_SOLICITOR_NAME).build());
-        applicant.setSolicitorRepresented(YesOrNo.YES);
-        final var caseData = CaseData.builder().applicant2(applicant).build();
+        caseData.setApplicationType(JOINT_APPLICATION);
         caseData.setConditionalOrder(ConditionalOrder.builder()
             .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
             .build());
 
         when(commonContent.basicTemplateVars(caseData, 1234567890123456L)).thenReturn(getBasicTemplateVars());
 
-        conditionalOrderPendingReminderNotification.sendToApplicant2Solicitor(caseData, 1234567890123456L);
+        awaitingConditionalOrderReminderNotification.sendToApplicant2Solicitor(caseData, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_SOLICITOR_EMAIL),
@@ -232,5 +172,21 @@ class ConditionalOrderPendingReminderNotificationTest {
             )),
             eq(ENGLISH)
         );
+    }
+
+    @Test
+    void shouldNotSendNotificationToApplicant2SolicitorIfSoleApplication() {
+        final Applicant applicant = getApplicant();
+        applicant.setSolicitor(Solicitor.builder().email(TEST_SOLICITOR_EMAIL).name(TEST_SOLICITOR_NAME).build());
+        applicant.setSolicitorRepresented(YesOrNo.YES);
+        final var caseData = CaseData.builder().applicant2(applicant).build();
+        caseData.setApplicationType(SOLE_APPLICATION);
+        caseData.setConditionalOrder(ConditionalOrder.builder()
+            .conditionalOrderApplicant1Questions(getConditionalOrderQuestions())
+            .build());
+
+        awaitingConditionalOrderReminderNotification.sendToApplicant2Solicitor(caseData, 1234567890123456L);
+
+        verifyNoInteractions(notificationService);
     }
 }
